getGameVariable fixed in Java, state's gameVariables member is now DoubleTensor in Lua
Former-commit-id: 12010f60f76b59fd5f7ea897ef0613d96fa5e3e9

+++ src/lib_java/ViZDoomGameJava.h
@@ -97,7 +97,7 @@ JNI_EXPORT(jint, getModeNative);
 JNI_EXPORT(void, setMode, jobject);
 JNI_EXPORT(jint, getTicrate);
 JNI_EXPORT(void, setTicrate, jint);
-JNI_EXPORT(jint, getGameVariable, jobject);
+JNI_EXPORT(jdouble, getGameVariable, jobject);
 JNI_EXPORT(jdouble, getLivingReward);
 JNI_EXPORT(void, setLivingReward, jdouble);
 JNI_EXPORT(jdouble, getDeathPenalty);
